+++ test/formatters/installers/goimports.bash
@@ -1,3 +1,3 @@
 apt-get install -y golang-go
-go get -u golang.org/x/tools/cmd/goimports
+go install golang.org/x/tools/cmd/goimports@latest
 cp -L "$HOME/go/bin/goimports" /usr/local/bin/